services/revenue-engine/requirements.txt
fastapi
uvicorn[standard]
scipy
numpy
pandas
pydantic
python-dotenv
httpx
